geotiff tests: consolidate sidecar cluster (#2436)
Fold the six top-level sidecar files into a new
integration/test_sidecar.py, one Section per source issue:

- test_sidecar_ovr_2112.py
- test_sidecar_own_geokeys_2315.py
- test_sidecar_max_cloud_bytes_2121.py
- test_sidecar_bad_does_not_break_base_2416.py
- test_remote_sidecar_chunked_2239.py
- test_remote_sidecar_byte_order_2314.py

Colliding module helpers are namespaced per source issue and GPU gating
moves to the shared requires_gpu marker. The golden_corpus fixture path
resolves against the tests directory now that the file lives one level
deeper. No assertion changed; 64 test functions (83 collected) match the
pre-consolidation total.

test_gpu_sidecar_georef_parity_2324.py imported _write_pair from the
own-geokeys file; the import now points at the consolidated module. The
release-gate checklist and geotiff docs cite the new path so the
checklist-parity gate stays green.

Sub-PR A of issue #2436 (cluster 12, long-tail epic #2424). Tests-only.
